Move all users views from home to users
